[Hackability Refactor] Easy Cheap Moves: eval, gguf_loader, quantize, model_dist
